code/requirements.txt
tensorflow==2.11.0
keras==2.11.0
opencv-python==4.7.0.72
Pillow==9.5.0